Initial spike of test resources support
This commit introduces "test resources" support for Micronaut.
Test resources work by implementing the `TestResourcesResolver`
interface, which consists of a couple methods:

- `listProperties` returns the list of properties that the
resolver is able to answer
- `resolve` is responsible for computing the value of those
properties

There are different implementations:

- in `core`, we implement the base logic which is to lazily
resolve properties. It works with a combination of a property
source loader and a property expression loader. The property
source loader is responsible for listing the properties which
can be resolved, while the loader is responsible for answering
lazily.
- in `embedded`, we provide an implementation which uses service
loading to load providers. Adding this to an application classpath,
then "property loaders" can be added just by implementing the
`TestResourceResolver` interface and declaring the service in
the appropriate service file.
- in `client`, we provide a lightweight implementation of a resolver
which delegates resolution to an external service (a proxy). This
would typically be used when the build tools inject resolvers, so
that the service responsible for spawning test containers live in
its own process, isolated from user code.

Note that currently, there's nothing really specific to "test"
resources: it's a generic mechanism which lets Micronaut application
resolve properties when they are missing. However, because it implies
service loading or a client, it would have an impact on startup times,
so should be limited to development mode or testing.