Convert the jdbc-mysql module into a generic JDBC module
This makes it easier to write other JDBC specific containers, at the
cost of putting on classpath more support libraries which are not
required by the user.

All the different JDBC test resource providers can provide the
value of the "datasources.default.url" property. Therefore, in order
to choose the one which is going to answer (and spawn the container),
we proceed with filtering, by looking at the "dialect" property. Only
the provider with the corresponding dialect would answer.