Register all containers for shutdown
This commit introduces a registry of all created containers:
because we only know when a container is started, but that it can
be re-used if, for example, another test requests the exact same
container, we register them in a static field.

A facility is provided to that all containers can be closed.