Add support for test resource resolvers which work on multiple properties
For example, a datasource test resource resolver may need to resolve
the "default" datasource, but also extra datasources declared by the
user.

For this purpose, a test resource resolver can now declare the
required property entries it needs. In the example above, a test
resource resolver would declare it needs the `datasources` entries.
This information is used _before_ it returns the list of keys it
is able to resolve: given the list of datasources, it's now capable
of generating the list of properties that it can resolve.