Add ability to define custom scopes
Previously to this commit, the only possible lifecycle for resources
was the JVM scope: when the JVM is shutdown, containers are. Now,
it's possible for a test resource provider to bind resources to
a particular scope.

Scopes are identified with a dotted string, which matches a hierarchy.
For example, scope `a.b.c` has `a.b` as a parent, which has `a` as
a parent, itself inheriting from the root scope. Closing a scope
closes all resources of the scope and children scopes.