Add classpath inference module
The goal of this module is to infer the server classpath, given
the application classpath. Therefore, instead of having to add
all possible test resources modules on classpath (and potentially
adding modules which are in conflict with each other), this module
will infer which modules to add to the server classpath given
what the application actually uses.

For example, if Micronaut Data is present and that a MySQL driver
is also present, then the `test-resources-jdbc-mysql` module will
be inferred, making it possible to spawn a MySQL container.