Add support for R2DBC
The R2DBC support is a bit special in the sense that it can work
in standalone mode, in which case it will first spawn a JDBC container
_then_ bridge it to R2DBC, or it can re-use an existing JDBC container.

This is useful in case a project uses Flyway, for example, which will
operate in JDBC mode, not R2DBC.

For this to happen, the code had to be refactored so that we can determine
what containers "resolved what property". For example, we needed a way
to figure out that a MySQL container was associated to the resolution
of the `datasources.default.url` property, so that if we try to resolve
`r2dbc.datasources.default.url`, then we can first search for a container
which supplied `datasources.default.url`. Note that this assumes that
the R2DBC and JDBC datasources are "mirrored", in the sense that they use
the same names. If not, strange things may happen.