Add shared utilities to start and stop a test resources server
This commit introduces a build tool API for handling the lifecycle
of test resources server. It provides an abstraction over the startup
and closing of the test server, handles writing of the required
client properties file, and handles the logic of explicit port handling.

Testing doesn't check the "explicit port" case yet, because testing
for explicit port is complicated in terms of mocking.