Optimize server with Micronaut AOT (#24)
* Run Micronaut AOT on the test resources server

Because we want it to start as fast as possible, this commit
performs AOT optimization of the server jar. This is the jar
which is going to be published instead of the regular one.

* Upgrade Gradle to 7.5-rc-2

This works around a corrupt serialization bug when resolving
dependencies after upgrading to Micronaut 3.5.1.

* Use a disambiguation rule

This should fix code coverage failing because it cannot select the right
variant.