Pass test resources configuration to create method
This commit fixes an inconsistency in the API: for generic test container
providers, the `create` method was only getting the explicit list of
requested properties, but not the test resources configuration map.

This forced implementors which need additional test resources configuration
to explicitly require those properties when they are actually already
available directly.