Align MsSQL version with TestContainers in starter
In the starter project, if a user pulls in the TestContainers feature, they get the URL and EULA for `2019-CU4-ubuntu-16.04`

https://github.com/micronaut-projects/micronaut-starter/blob/f7902691bb355c8a4f923dce086b113c846b8346/starter-core/src/main/java/io/micronaut/starter/feature/database/TestContainers.java#L64
https://github.com/micronaut-projects/micronaut-starter/blob/f7902691bb355c8a4f923dce086b113c846b8346/starter-core/src/main/java/io/micronaut/starter/feature/database/TestContainers.java#L148

In the test-resources spike branch of starter, we're seeing weird results if the user now pulls in TestContainers as the feature
writes the EULA for `2019-CU4-ubuntu-16.04` but then the version TestResources starts is `2017-CU2`.

This PR aligns these versions so they are at least the same.

However, it raises 2 questions:

1. Should we upgrade to a later CU?
2. What should actually happen if a user pulls in TestResources?

For 2, they have TestResources implicitly so they get no jdbc/r2dbc URL, but then we're writing EULA files. We need to think
about what a user would expect from pulling in both of these features.