Avoid spawning an Oracle container when Oracle cloud is used (#110)
In production, when Oracle ATP is used, we wouldn't set the
datasource URL, but instead set the `ocid` property (and possibly
others). This confuses test resources, because if a property is
missing, then it would try to resolve it. Which is what happens here:
even if the `url` property is effectively unused when the cloud
environment is detected, the property was effectively resolved.

The workaround is to try to see if the `ocid` property is set, in
which case the resolvers wouldn't do anything.