Fix property resolving when client returns a 404
This commit fixes handling of the resolution result whenever the client
throws an error, which is the case if the resolver returns an empty
optional. This improves the error message on the client side.

It also improves the PostgreSQL providers so that they support both
`postgres` and `postgresql` as `db-type`, since it's a common mistake.

Fixes #129