Support reading server connection details from system properties
This commit adds the ability for the client to get the server
connection details from system properties. Before, the only
possible solution was to have a `test-resources.properties`
file on classpath.

Because the client and the configuration had to be injected
via classpath, it meant that a change to the server configuration
would typically invalidate classpath and cause tasks to be
out-of-date in Gradle (see https://github.com/micronaut-projects/micronaut-gradle-plugin/issues/579)

The new option makes it possible to read the server
configuration from system properties if the file isn't
found on classpath. This preserves backwards compatibility
while offering a convenient way to pass in the parameters.

Note that we are not using environment variables, because
there would be no option to pass them via Gradle.