Fix flat jar being deleted when directories change
This commit fixes an issue with the logic when the AppCDS classpath
changes: we have logic to "snapshot" the classpath in order to
detect changes. We also have logic to make sure that we build a
"flat jar" which contains the contents of all directories on
classpath (because AppCDS only supports jars on classpath).

When a directory on classpath changed, for example if we made a
change to `src/testResources/main/some/Resource.java`, then we
would properly detect the change, but we accidentally deleted
the newly generated flat jar.

Fixes #148