Use a fat jar for the test resources service (#165)
* Use the Micronaut Platform BOM

Until a version of the Gradle plugin compatible with Micronaut 4
is out.

* Make use of a fat jar

This commit changes how the server is built, in order to use a fat
jar for the server instead of a regular jar. The reason is that the
build tools need a way to provide the server with default versions
for the libraries that the user defines in their build scripts.

For this, we want to inject the Micronaut Platform BOM on the
server classpath. But that BOM would be the BOM that for the
version of Micronaut that the user defines, it is **not** the
one for the version that the test resources service uses itself.

By using a regular jar, we have the problem that the jar has
dependencies, including the BOM, so we'd get a conflict. By
using a fat jar, we don't have any dependencies anymore for
the test resources service, so adding the user Micronaut BOM
has no risk of breaking the test resources service.