Enable parallel build
This commit enables parallel building of projects. While I was investigating
support for Gradle's configuration cache, I noticed that if the cache is
enabled, then projects are built in parallel _even if parallel is false_
or that `--no-parallel` is used. While this may be a bug in Gradle, there
is value in supporting parallel builds.

The reason it wasn't done before is that test resources tests are resource
intensive: most of them require docker containers. In addition, the tests
make assertions on the number of containers which are started. If 2 projects
use the same container kind in parallel (say, Kafka), then the assertions
would fail, as seen in https://ge.micronaut.io/s/sl32a7lohhmio/tests/:micronaut-test-resources-localstack-s3:test/io.micronaut.testresources.localstack.s3.LocalStackS3Test/automatically%20starts%20an%20S3%20container?top-execution=1

Therefore, we introduce a parallel gate service, which allows declaring
resources in use, and combine them. For example, a project can declare
that it uses MySQL. We only allow a single database type to be used in
parallel (that is to say that a single test using MySQL can run at a
point in time), but we do allow multiple tests using different dbs to
run in parallel. However, we also limit the number of concurrent db
tests. In addition, we can cap the total number of parallel tests.