Only use kebab case for resolvable properties
This commit fixes the test resource resolvers which were returning
properties using camel case. In order to make sure new resolvers,
or resolvers implemented by users, only make use of kebab case,
we added a check in the embedded resolver: this resolver is the
one which is used to collect all resolvers from classpath, and
is used in both the test resources service and in embedded mode.

Fixes #183