Implement test resources property provider factory
This commit introduces a new annotation, `@TestResourceProperties`, which
can be applied on a test in order to tell that the test needs to resolve
one or more test resource properties _before_ the application context is
available.

For example, with:

```java
@MicronautTest
@TestResourcesProperties(
    value = "redis.url"
)
class SomeTest { ... }
```

Then the value of the `redis.url` property will automatically be
fetched from the test resources service and made available as if
it had been done in a `TestPropertyProvider`. The difference is
that this without this annotation, it was required to call the
test resources client directly in the provider, which was error
prone. In particular, the `properties` argument of the `resolve`
call are not easy to figure out.

This annotation is therefore a convention to avoid having to
call the client directly. However, in some cases it might be
necessary to read a property in order to compute a different
one which needs to be available in the test.

This can be done by adding a `providers` argument to the
annotation:

```java
@TestResourcesProperties(
    value = "rabbitmq.uri",
    providers = RabbitTest.RabbitMQProvider.class
)
class RabbitTest {
  // ...
      @ReflectiveAccess
      public static class RabbitMQProvider implements TestResourcesPropertyProvider {
          @Override
          public Map<String, String> provide(Map<String, Object> testProperties) {
              String uri = (String) testProperties.get("rabbitmq.uri");
              return Map.of(
                  "rabbitmq.servers.product-cluster.port", String.valueOf(URI.create(uri).getPort())
              );
          }
      }
}
```

The `TestResourcesPropertyProvider` type has access to all properties
which are already resolved at the moment it is called, which includes
the properties asked in `@TestResourcesProperties`.

This should make the implementation of https://github.com/micronaut-projects/micronaut-nats/pull/321
easier.