Add a `TestResourcesScope` annotation (#241)
By default, test resources are shared between all tests. This means,
in particular, that a single container will be spawned if multiple
tests require the same property. For example, we would have a single
MySQL container for the whole test suite, which makes testing
significantly faster, avoiding containers to be started and shutdown
in each test.

However, in some situations it may be required to isolate tests
from each other, and a test may want to work with its own, isolated
container.

This commit introduces a new `TestResourcesScope` annotation which
can be put added at a _test class_ level, which defines the scope
of the test resources used in that test. A scope can be shared
by multiple tests, and you can see the general case (all tests
share the same resources) as the "root scope". A scope is closed
whenever the last test which works in that scope is finished.

Before, doing this required assigning a test property to the
test and manually counting the number of tests which are executed,
then calling the test resources client directly when all tests
are executed, which is both error prone and imprecise (in case
only a subset of the tests are executed).

Because of implementation limitations, the way the scope is
recognized is by using a thread local, which has a couple
consequences:

1. it only works properly if tests are executed in a _single_
JVM.
2. tests which use custom threads and execute the application
in that thread wouldn't see the scope and have to rely on
using the client