Make sure we can't use MariaDB instead of MySQL (#252)
or the other way around. Currently, if both the MariaDB and MySQL
test resources providers are present, then because they share the
same dialect, it is possible that we select one provider or the
other randomly, based on which comes first.

To fix this, we now make sure that if the `db-type` property is
set, then we only return a container if it is actually capable
of handling that db type. Since this test comes first, before
the dialect, it has higher precedence and we wouldn't pick one
container for the other.