Fix classpath leak of test resources modules (#266)
* Fix classpath leak of test resources modules

For a few versions already, the test resources service has been
shipped as a fat jar. This was done in order to make it possible
to apply the Micronaut BOM to the server so that the user can
add dependencies like MySQL drivers and so without having to
specify a version.

However, when we did the change, we didn't change how the _modules_
were built. Typically, the `testcontainers` module would have
a dependency on `micronaut-inject`, which was then injected
into the runtime classpath of the server. If all versions are
aligned, this isn't an issue. However, as soon as _one_ version
isn't, then things blow up.

This commit fixes the problem by making sure that the dependencies
on Micronaut for the modules are _compile only_ dependencies, since
their runtime is going to be provided by the server.

This will also require changes in the Micronaut Gradle plugin, since
it offers an option to have a specific source set for building
Micronaut Test Resources extensions, and the same change has to be
done there.

This fixes a weird bug which was happening when upgrading the Gradle
plugin to Micronaut Test Resources 2.0.0-M6, where it was complaining
about a missing snakeyml dependency, when the server effectively
doesn't need it (as verified when running the fat jar directly on
command line).

* Fix test of client