Fix YamlPropertySourceLoader being added on classpath (#267)
The previous fix (#266) wasn't sufficient, as when you don't use
any YAML configuration file, the YAML property source loader is
still added my Micronaut AOT to the list of services which are
"statically" loaded. This probably requires a fix in Micronaut AOT.
In between, this configuration makes sure that the loader isn't
added.