Implement binary protocol between client and server
This commit introduces a custom protocol for communicating
between the client and the server. Instead of using JSON, it now
uses a binary protocol. This removes the requirement of having
Micronaut JSON on the user classpath (on the client side) and
further reduces the number of dependencies of the server.

This is a major breaking change, and it's no longer possible
to directly use tools like cURL or httpie to talk to the server.

This is blocked by https://github.com/micronaut-projects/micronaut-core/issues/9331