Let the client factory find local configuration files
In normal operation mode, the server settings are passed to the
client via system properties. This lets the client figure out
on which server to connect. In particular, there may be more than
one test resources server running, and there can be shared servers
as well as namespaced servers.

However, in practice IDEs may not delegate to the build tool,
which means that the configuration is not passed to the client.
We want to make it possible to support the following scenario:

1. start a test resources service from command line via the
build tool, e.g `mvn mn:start-test-resources-service`
2. in the IDE, run tests or main application without delegating
to the build tool
3. tests should pass

It is currently working by accident for tests in Maven, because
the properties file is on classpath. It doesn't work for the
main application, though.

This commit removes the ability to search for the configuration
on classpath, which is unreliable, and replaces it with a
"conventional" lookup, which searches the configuration in that
order:

1. from system properties, the preferred way, configured by
build plugins
2. from a project local directory, `.micronaut/test-resources`
3. from the user home, under `.micronaut/test-resources`

While 1. is reliable, 2 and 3 are not, since they _cannot_
support all the scenarios that test resources support. In
particular, they don't support namespacing. The Gradle plugin
doesn't write its configuration into `.micronaut/test-resources`
either, since it is possible for a project to run several
test resources services concurrently.