Add support for dependencies between containers (#360)
Because of technical reasons, the support is currently limited to
generic containers: in this case, we know that the `properties`
map which is passed when resolving a property will always be
empty, because a generic container cannot declare such required
properties. Therefore, when a container is resolved, we can look
into the _other_ generic containers which are declared, and see
if one matches the required id. If we find one, then we can
trigger the resolution of the container by resolving one of
the properties it provides (any property works).

This feature does _not_ rely on the `dependsOn` feature of
Testcontainers for the very same reason: the resolver doesn't
have access to other containers which are resolved, nor does
it know how to create one at the point in time the resolution
is triggerred. Therefore, we have no `Container` instance to
pass to `dependsOn`.

Fixes #147