Add property to disable some test resources resolvers (#401)
* Add property to disable some test resources resolvers

This commit introduces a property to disable a test resolver.
It was implemented in a backwards compatible fashion, by introducing
a new subtype of resolver. Resolvers which can be disabled need
to implement the `ToggableTestResourcesResolver` interface and
provide a name.

This name is used at resolution time, to figure out a test
resources property. For example, if the name of the resolver
is foo, then at resolution time, we'd look into the test
resources configuration for the property:

test-resources.foo.enabled

and ignore the resolver if that property is `false`. This flag
should only be used for debugging purposes.

Fixes #399

* Bump version