Introduce test resources control panel
This module introduces an _optional_ control panel, which can be
typically enabled by adding the module in the "additionalModules"
section of the build:

```
micronaut {
   testResources {
      additionalModules.add("control-panel")
   }
}
```

This module provides information about the status of
the test resources service, such as the started
containers, or the properties which were resolved.

This can be used to diagnose errors, or to figure
out things like the database connection parameters
of the current container.