Improve observability of starting containers
This commit improves the observability of containers being started.
First, it changes how locking works, so that we permit starting
different containers concurrently, and also permit reading the
list of containers while another container is being started.
Previously, the fact of having _one_ container starting blocked
all other requests for containers (including listing), which
was not very practical.

In addition, it is now possible to see which containers are
being started.