Fix server not shutting down (#464)
It seems that in some cases, even if the default stop endpoint is
called, the server is not shutdown properly. This seems to happen
if some containers are started, but not in all cases. With this
change, we switch from using the `/stop` endpoint from management
to a custom one which times out after 5s, making sure the server
is killed.