Fix server startup utils (#478)
* Fix server startup utils

This commit fixes a bug where we would not wait for the test resources
service to be available when using an explicit port. Instead, the
utility was returning immediately, which caused the test resources
client to fail if it tried to connect before the service had a chance
to initialize. This problem can surface more easily when, during
startup, we perform tests like checking that Docker is available.

* Update test-resources-build-tools/src/main/java/io/micronaut/testresources/buildtools/ServerUtils.java

Co-authored-by: Álvaro Sánchez-Mariscal <alvaro.sanchez-mariscal@oracle.com>

---------

Co-authored-by: Álvaro Sánchez-Mariscal <alvaro.sanchez-mariscal@oracle.com>