Fix NPE when client is on classpath but not configured
Because of user errors (for example explicitly adding the test
resources client even if disabled) or simply because the test
resources extensions have the client as a transitive dependency,
it is possible to have the test resources client present on
classpath even if test resources are disabled. In that case,
the client wouldn't be configurable.

This commit fixes what the client returns in case there's no
test resources service configured, by returning reasonable
defaults instead of null.

Fixes #651