Fix test resources properties factory
This commit fixes the `TestResourcesPropertiesFactory`, which
incorrectly passed the `testResourcesConfig` map. Instead of
removing the `test-resources.` prefix, like the client expects,
it was passing the full property name which led to resolvers
being unable to resolve properties.

While this commit fixes the problem at the source, in the
properties factory, a sanity cleanup has been introduced
in the controller, in order to cleanup such entries in case
they show up. This could be useful if a program uses the
client directly but forgets to do the same (in the "normal"
use case, the properties are passed "naturally" without the
prefix because that's how the config API works).

Fixes #663