Use lenient docker checks
This commit changes how docker checks are done, so that if the docker
environment is detected, then nothing changes, but if it is not, then
we would still try to start a container, because it seems that in
some cases, the docker environment is present but not properly detected,
or detected too late. In this case, we will now display a warning,
and let the docker check re-execute for the next container.